// lib/interfaces/IEmployee.ts
export interface EmployeeModel {
  avatar: string;
  email: string;
  salary: string;
  date: string;
  status: string;
  firstname: string;
  lastname: string;
}

export interface EmployeeData {
  _id?: string;
  name: string;
  avatar: string;
  email: string;
  salary: string;
  date: string;
  status: string;
}